Switch to stdlib go/types and /x/tools/go/loader
go/types has moved to the standard library, and honnef.co/go/importer
has been removed.

This patch is slightly inefficient, as it typechecks packages twice. We
could fix it, however we plan to deprecate this program in favour of
using guru, so there's no point in putting too much effort in.